fix: REWORK round 2 — build, file-size cap, retrieval-perf workflow
1. Fix build failure: add rerank_candidate_cap, multi_query_max,
   llm_token_ceiling_per_tenant to Config initializer in
   integration_hybrid_tenant_isolation_tests.rs (same fix already
   applied to src/jobs/tests.rs in da83de3).

2. Fix 600-line cap: extract clamp_rerank_candidates and
   llm_budget_allows to a new cost_ceilings.rs submodule (45 lines);
   re-export from search.rs via pub(crate) use. search.rs 631→585 lines.

3. Fix retrieval-perf.yml: add uses: ./.github/actions/setup-rdkafka
   step (installs libcurl4-openssl-dev and friends needed by rdkafka-sys).

Production code (AC1–AC8) unchanged from round 1 approval.
