+++ .github/workflows/retrieval-perf.yml
@@ -56,6 +56,8 @@ jobs:
 
       - uses: Swatinem/rust-cache@v2
 
+      - uses: ./.github/actions/setup-rdkafka
+
       - name: Run retrieval latency budget tests
         run: |
           cargo test -p control-api \
